bigquery_sync/requirements.txt
google-cloud-bigquery==2.34.4
psycopg2-binary==2.9.5
numpy<2.0.0  # Pin NumPy to version 1.x

protobuf<3.18.0